Review form test written and passing
