+++ app/controllers/reviews_controller.rb
@@ -1,6 +1,12 @@
 class ReviewsController < ApplicationController
 
   def new
+    @item = Item.find(params[:item_id])
+    @review = Review.new
+  end
+
+  def create
+
   end
 
 end
