+++ app/views/profile/orders/_order_item.html.erb
@@ -9,6 +9,8 @@
     <p>Quantity: <%= oitem.quantity %></p>
     <p>Subtotal: <%= number_to_currency(oitem.subtotal) %></p>
     <p>Fulfilled: <%= oitem.fulfilled ? 'Yes' : 'No' %></p>
+    <% if oitem.order.status == "completed" %>
     <p><%= link_to "Review", new_item_review_path(oitem.item) %></p>
+    <% end %>
   </div>
 </div>
