+++ app/views/reviews/new.html.erb
@@ -1,0 +1,9 @@
+<%= form_for [@item, @review] do |f|  %>
+  <p><%= f.label :title %>
+  <%= f.text_field :title %></p>
+  <p><%= f.label :description %>
+  <%= f.text_field :description%></p>
+  <p><%= f.label :rating %>
+  <%= f.number_field :rating%></p>
+  <%= f.submit 'Submit' %>
+<% end %>
